backend/requirements.txt
fastapi
uvicorn
paddleocr
pdfplumber
pdf2image
pillow
python-dotenv
httpx
supabase
